# perf-simple-query: add mean and standard deviation stats