@bhalevy 2024-06-25
@scylladb/scylla-maint please help review/merge this change.